use new div on each render
+ updates to tests & types
