+++ browser/index.js
@@ -1,17 +1,14 @@
 /* global window */
 /* global ReadableStream */
-
 import { render } from 'lit-html';
 export * from 'lit-html';
-
-const RENDER_ELEMENT = window.document.createElement('div');
 
 /**
  * lit-html-server API parity
  * @param {import('lit-html').TemplateResult} result
  * @returns {ReadableStream}
  */
-export function renderToStream(result) {
+export function renderToStream(result /*options*/) {
   if (typeof ReadableStream === 'undefined') {
     throw Error('ReadableStream not supported on this platform');
   }
@@ -30,7 +27,7 @@ export function renderToStream(result) {
  * @param {import('lit-html').TemplateResult} result
  * @returns {Promise<Uint8Array>}
  */
-export async function renderToBuffer(result) {
+export async function renderToBuffer(result /*options*/) {
   if (typeof TextEncoder === 'undefined') {
     throw Error('TextEncoder not supported on this platform');
   }
@@ -42,8 +39,9 @@ export async function renderToBuffer(result) {
  * @param {import('lit-html').TemplateResult} result
  * @returns {Promise<string>} html string
  */
-export function renderToString(result) {
-  render(result, RENDER_ELEMENT);
-  const htmlString = RENDER_ELEMENT.innerHTML;
+export function renderToString(result /*options*/) {
+  const renderElement = window.document.createElement('div');
+  render(result, renderElement);
+  const htmlString = renderElement.innerHTML;
   return Promise.resolve(htmlString);
 }
