+++ tests/index.html
@@ -18,7 +18,7 @@
         mocha.run();
       }
     </script>
-    <script type="module" src="./tests.js"></script>
+    <script type="module" src="./test-render.js"></script>
 
 
     </head>
